+++ NEWS.rst
@@ -6,7 +6,9 @@ of every change, see the Git log.
 
 Latest
 ------
-* tbd
+* Major: Removed C++ Standard from mkspecs. The C++ Standard can now be
+  specified with the ``conf.set_cxx_std`` method. If no C++ Standard is
+  set explicitly, then the default is whatever the compiler uses by default.
 
 4.44.0
 ------